+++ deltachat-jsonrpc/src/api/types/mod.rs
@@ -6,6 +6,7 @@ pub mod account;
 pub mod chat;
 pub mod chat_list;
 pub mod contact;
+pub mod location;
 pub mod message;
 pub mod provider_info;
 pub mod webxdc;